typecase to size_t in assert

+++ Include/cpython/pyatomic_msc.h
@@ -1164,7 +1164,7 @@ _Py_atomic_memcpy_ptr_store_relaxed(void *dest, void *src, size_t n)
 {
     // assert(((uintptr_t)dest & (sizeof (void *) - 1)) == 0);
     // assert(((uintptr_t)src & (sizeof (void *) - 1)) == 0);
-    assert(n % sizeof(void *) == 0);
+    assert(n % (size_t)sizeof(void *) == 0);
 
     // memcpy(dest, src, n);
 
@@ -1187,7 +1187,7 @@ _Py_atomic_memmove_ptr_store_relaxed(void *dest, void *src, size_t n)
 {
     // assert(((uintptr_t)dest & (sizeof (void *) - 1)) == 0);
     // assert(((uintptr_t)src & (sizeof (void *) - 1)) == 0);
-    assert(n % sizeof(void *) == 0);
+    assert(n % (size_t)sizeof(void *) == 0);
 
     // memmove(dest, src, n);
